adversarial_cloaker/requirements.txt
torch
torchvision
numpy<2
matplotlib
pillow
tqdm
seaborn
scikit-learn
lpips